packages/llama-agents-appserver/pyproject.toml
[build-system]
requires = ["uv_build>=0.7.20,<0.8.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
  "pytest>=8.4.1",
  "pytest-asyncio>=0.25.3",
  "respx>=0.22.0",
  "pytest-xdist>=3.8.0",
  "ty>=0.0.15",
  "ruff>=0.12.9"
]

[project]
name = "llama-agents-appserver"
version = "0.10.2"
description = "Application server components for LlamaDeploy"
readme = "README.md"
license = {text = "MIT"}
authors = [
  {name = "Massimiliano Pippi", email = "mpippi@gmail.com"},
  {name = "Adrian Lyjak", email = "adrianlyjak@gmail.com"}
]
requires-python = ">=3.10, <4"
dependencies = [
  "llama-index-workflows>=2.16.0,<3.0.0",
  "llama-agents-server>=0.2.2,<0.3.0",
  "pydantic-settings>=2.10.1",
  "fastapi>=0.100.0",
  "websockets>=12.0",
  "llama-agents-core>=0.5.0",
  "httpx>=0.24.0,<1.0.0",
  "prometheus-fastapi-instrumentator>=7.1.0",
  "packaging>=25.0",
  "structlog>=25.4.0",
  "rich>=14.1.0",
  "pyyaml>=6.0.2",
  "watchfiles>=1.1.0",
  "uvicorn>=0.35.0",
  "typing-extensions>=4.15.0 ; python_full_version < '3.12'"
]

[tool.uv.build-backend]
module-name = ["llama_agents.appserver", "llama_deploy.appserver"]
namespace = true

[tool.uv.sources]
llama-agents-core = {workspace = true}
